requirements.txt
lxml==4.5.0
xmldiff==2.4
google-cloud-bigquery==1.23.1
pyyaml==5.3.1
pandas==1.0.3
pytest==5.4.1
pycodestyle==2.6.0
pytest-cov==2.8.1